+++ rotasim/__init__.py
@@ -1,2 +1,3 @@
 from .version import __version__, __versiondate__
 from .rotasim import *
+from .rotasim_genetics import *
