+++ rotasim/rotasim.py
@@ -3,7 +3,7 @@ Rotasim model
 
 Usage:
     import rotasim as rs
-    sim = rs.RotaABM()
+    sim = rs.RotaSim()
     sim.run()
 
 TODO:
@@ -21,625 +21,69 @@ import random as rnd
 import numpy as np
 import itertools
 import sciris as sc
+import starsim as ss
+from . import rotasim_genetics as rg
 
 # Define age bins and labels
 age_bins = [2/12, 4/12, 6/12, 12/12, 24/12, 36/12, 48/12, 60/12, 100]
 age_distribution = [0.02, 0.02, 0.02, 0.02, 0.02, 0.02, 0.02, 0.02, 0.84]                  # needs to be changed to fit the site-specific population
 age_labels = ['0-2', '2-4', '4-6', '6-12', '12-24', '24-36', '36-48', '48-60', '60+']
 
-__all__ = ['HostPop', 'Host', 'Pathogen', 'Sim']
+__all__ = ['RotaSim']
+
+#
+# ### Host classes
+# class HostPop(ss.People):
+#     """
+#     A class to hold all the hosts
+#     """
+#     def __init__(self, N, sim):
+#         self.hosts = [Host(i, sim) for i in range(N)]
+#         self.bdays = [h.bday for h in self.hosts]
+#         self.ids = [h.id for h in self.hosts]
+#         # self.vaccinated_hosts = [h for h in self.hosts if h.vaccine is not None]
+#         return
+#
+#     def __repr__(self):
+#         return f'HostPop(N={len(self)})'
+#
+#     def __iter__(self):
+#         return self.hosts.__iter__()
+#
+#     def __len__(self):
+#         return len(self.hosts)
+#
+#     def __getitem__(self, key):
+#         return self.hosts[key]
+#
+#     def append(self, host):
+#         self.hosts.append(host)
+#         self.bdays.append(host.bday)
+#         return
+#
+#     def remove(self, value):
+#         index = self.hosts.index(value)
+#         del self.hosts[index]
+#         del self.bdays[index]
+#         return
 
 
-### Host classes
-class HostPop:
-    """
-    A class to hold all the hosts
-    """
-    def __init__(self, N, sim):
-        self.hosts = [Host(i, sim) for i in range(N)]
-        self.bdays = [h.bday for h in self.hosts]
-        self.ids = [h.id for h in self.hosts]
-        # self.vaccinated_hosts = [h for h in self.hosts if h.vaccine is not None]
-        return
 
-    def __repr__(self):
-        return f'HostPop(N={len(self)})'
 
-    def __iter__(self):
-        return self.hosts.__iter__()
 
-    def __len__(self):
-        return len(self.hosts)
-
-    def __getitem__(self, key):
-        return self.hosts[key]
-
-    def append(self, host):
-        self.hosts.append(host)
-        self.bdays.append(host.bday)
-        return
-
-    def remove(self, value):
-        index = self.hosts.index(value)
-        del self.hosts[index]
-        del self.bdays[index]
-        return
-
-
-class Host(sc.prettyobj):
-    """
-    A rotavirus host
-    """
-    def __init__(self, host_id, sim):
-        self.sim = sim
-        self.id = host_id
-        self.bday = self.t - self.get_random_age()
-        self.immunity = {} # set of strains the host is immune to
-        self.vaccine = None
-        self.infecting_pathogen = []
-        self.prior_infections = 0
-        self.prior_vaccinations = []
-        self.infections_with_vaccination = []
-        self.infections_without_vaccination = []
-        self.is_immune_flag = False
-        self.oldest_infection = np.nan
-        return
-
-    @property
-    def t(self):
-        return self.sim.t
-
-    @property
-    def numAgSegments(self):
-        return self.sim.numAgSegments
-
-    @staticmethod
-    def get_random_age():
-        # pick a age bin
-        random_age_bin = np.random.choice(list(range(len(age_bins))), p=age_distribution)
-        # generate a random age in the bin
-        if random_age_bin > 0:
-            min_age = age_bins[random_age_bin-1]
-        else:
-            min_age = 0
-        max_age = age_bins[random_age_bin]
-        return rnd.uniform(min_age, max_age)
-
-    def get_age_category(self):
-        # Bin the age into categories
-        for i in range(len(age_bins)):
-            if self.t - self.bday < age_bins[i]:
-                return age_labels[i]
-        return age_labels[-1]
-
-    def get_oldest_current_infection(self):
-        max_infection_times = max([self.t - p.creation_time for p in self.infecting_pathogen])
-        return max_infection_times
-
-    def compute_combinations(self):
-        seg_combinations = []
-
-        # We want to only reassort the GP types
-        # Assumes that antigenic segments are at the start
-        for i in range(self.numAgSegments):
-            availableVariants = set([])
-            for j in self.infecting_pathogen:
-                availableVariants.add((j.strain[i]))
-            seg_combinations.append(availableVariants)
-
-        # compute the parental strains
-        parantal_strains = [j.strain[:self.numAgSegments] for j in self.infecting_pathogen]
-
-        # Itertools product returns all possible combinations
-        # We are only interested in strain combinations that are reassortants of the parental strains
-        # We need to skip all existing combinations from the parents
-        # Ex: (1, 1, 2, 2) and (2, 2, 1, 1) should not create (1, 1, 1, 1) as a possible reassortant if only the antigenic parts reassort
-
-        # below block is for reassorting antigenic segments only
-        all_antigenic_combinations = [i for i in itertools.product(*seg_combinations) if i not in parantal_strains]
-        all_nonantigenic_combinations = [j.strain[self.numAgSegments:] for j in self.infecting_pathogen]
-        all_strains = set([(i[0] + i[1]) for i in itertools.product(all_antigenic_combinations, all_nonantigenic_combinations)])
-        all_pathogens = [Pathogen(self.sim, True, self.t, host = self, strain=tuple(i)) for i in all_strains]
-
-        return all_pathogens
-
-    def isInfected(self):
-        return len(self.infecting_pathogen) != 0
-
-    def recover(self,strain_counts):
-        # We will use the pathogen creation time to count the number of infections
-        creation_times = set()
-        for path in self.infecting_pathogen:
-            strain = path.strain
-            if not path.is_reassortant:
-                strain_counts[strain] -= 1
-                creation_times.add(path.creation_time)
-                self.immunity[strain] = self.t
-                self.is_immune_flag = True
-                if np.isnan(self.oldest_infection):
-                    self.oldest_infection = self.t
-        self.prior_infections += len(creation_times)
-        self.infecting_pathogen = []
-        self.possibleCombinations = []
-
-    def vaccinate(self, vaccinated_strain):
-        if len(self.prior_vaccinations) == 0:
-            self.prior_vaccinations.append(vaccinated_strain)
-            self.vaccine = ([vaccinated_strain], self.t, 1)
-        else:
-            self.prior_vaccinations.append(vaccinated_strain)
-            self.vaccine = ([vaccinated_strain], self.t, 2)
-
-    def is_vaccine_immune(self, infecting_strain):
-        # Effectiveness of the vaccination depends on the number of doses
-        if self.vaccine[2] == 1:
-            ve_i_rates = self.sim.vaccine_efficacy_i_d1
-        elif self.vaccine[2] == 2:
-            ve_i_rates = self.sim.vaccine_efficacy_i_d2
-        else:
-            raise NotImplementedError(f"Unsupported vaccine dose: {self.vaccine[2]}")
-
-        # Vaccine strain only contains the antigenic parts
-        vaccine_strain = self.vaccine[0]
-        vaccine_hypothesis = self.sim.vaccine_hypothesis
-
-        if vaccine_hypothesis == 0:
-            return False
-        if vaccine_hypothesis == 1:
-            if infecting_strain[:self.numAgSegments] in vaccine_strain:
-                if rnd.random() < ve_i_rates[PathogenMatch.HOMOTYPIC]:
-                    return True
-                else:
-                    return False
-        elif vaccine_hypothesis == 2:
-            if infecting_strain[:self.numAgSegments] in vaccine_strain:
-                if rnd.random() < ve_i_rates[PathogenMatch.HOMOTYPIC]:
-                    return True
-                else:
-                    return False
-            strains_match = False
-            for i in range(self.numAgSegments):
-                immune_genotypes = [strain[i] for strain in vaccine_strain]
-                if infecting_strain[i] in immune_genotypes:
-                    strains_match = True
-            if strains_match:
-                if rnd.random() < ve_i_rates[PathogenMatch.PARTIAL_HETERO]:
-                    return True
-            else:
-                return False
-        # used below hypothesis for the analysis in the report
-        elif vaccine_hypothesis == 3:
-            if infecting_strain[:self.numAgSegments] in vaccine_strain:
-                if rnd.random() < ve_i_rates[PathogenMatch.HOMOTYPIC]:
-                    return True
-                else:
-                    return False
-            strains_match = False
-            for i in range(self.numAgSegments):
-                immune_genotypes = [strain[i] for strain in vaccine_strain]
-                if infecting_strain[i] in immune_genotypes:
-                    strains_match = True
-            if strains_match:
-                if rnd.random() < ve_i_rates[PathogenMatch.PARTIAL_HETERO]:
-                    return True
-            else:
-                if rnd.random() < ve_i_rates[PathogenMatch.COMPLETE_HETERO]:
-                    return True
-                else:
-                    return False
-        else:
-            raise NotImplementedError("Unsupported vaccine hypothesis")
-
-    def can_variant_infect_host(self, infecting_strain, current_infections):
-        numAgSegments = self.numAgSegments
-        immunity_hypothesis = self.sim.immunity_hypothesis
-        partial_cross_immunity_rate = self.sim.partial_cross_immunity_rate
-        complete_heterotypic_immunity_rate = self.sim.complete_heterotypic_immunity_rate
-        homotypic_immunity_rate = self.sim.homotypic_immunity_rate
-
-        if self.vaccine is not None and self.is_vaccine_immune(infecting_strain):
-            return False
-
-        current_infecting_strains = (i.strain[:numAgSegments] for i in current_infections)
-        if infecting_strain[:numAgSegments] in current_infecting_strains:
-            return False
-
-        def is_completely_immune():
-            immune_strains = (s[:numAgSegments] for s in self.immunity.keys())
-            return infecting_strain[:numAgSegments] in immune_strains
-
-        def has_shared_genotype():
-            for i in range(numAgSegments):
-                immune_genotypes = (strain[i] for strain in self.immunity.keys())
-                if infecting_strain[i] in immune_genotypes:
-                    return True
-            return False
-
-        if immunity_hypothesis == 1:
-            if is_completely_immune():
-                return False
-            return True
-
-        elif immunity_hypothesis == 2:
-            if has_shared_genotype():
-                return False
-            return True
-
-        elif immunity_hypothesis in [3, 4, 7, 8, 9, 10]:
-            if is_completely_immune():
-                if immunity_hypothesis in [7, 8, 9, 10]:
-                    if rnd.random() < homotypic_immunity_rate:
-                        return False
-                else:
-                    return False
-
-            if has_shared_genotype():
-                if rnd.random() < partial_cross_immunity_rate:
-                    return False
-            elif immunity_hypothesis in [4, 7, 8, 9, 10]:
-                if rnd.random() < complete_heterotypic_immunity_rate:
-                    return False
-            return True
-
-        elif immunity_hypothesis == 5:
-            immune_ptypes = (strain[1] for strain in self.immunity.keys())
-            return infecting_strain[1] not in immune_ptypes
-
-        elif immunity_hypothesis == 6:
-            immune_ptypes = (strain[0] for strain in self.immunity.keys())
-            return infecting_strain[0] not in immune_ptypes
-
-        else:
-            raise NotImplementedError(f"Immunity hypothesis {immunity_hypothesis} is not implemented")
-
-    def record_infection(self, new_p):
-        if len(self.prior_vaccinations) != 0:
-            vaccine_strain = self.prior_vaccinations[-1]
-            self.infections_with_vaccination.append((new_p, new_p.match(vaccine_strain)))
-        else:
-            self.infections_without_vaccination.append(new_p)
-
-    def infect_with_pathogen(self, pathogen_in, strain_counts):
-        """ This function returns a fitness value to a strain based on the hypothesis """
-        fitness = pathogen_in.get_fitness()
-
-        # e.g. fitness = 0.8 (there's a 80% chance the virus infecting a host)
-        if rnd.random() > fitness:
-            return False
-
-        # Probability of getting a severe decease depends on the number of previous infections and vaccination status of the host
-        severity_probability = self.sim.get_probability_of_severe(self.sim, pathogen_in, self.vaccine, self.prior_infections)
-        if rnd.random() < severity_probability:
-            severe = True
-        else:
-            severe = False
-
-        new_p = Pathogen(self.sim, False, self.t, host = self, strain=pathogen_in.strain, is_severe=severe)
-        self.infecting_pathogen.append(new_p)
-        self.record_infection(new_p)
-
-        strain_counts[new_p.strain] += 1
-
-        return True
-
-    def infect_with_reassortant(self, reassortant_virus):
-        self.infecting_pathogen.append(reassortant_virus)
-
-
-
-### Pathogen classes
-
-class PathogenMatch:
-    """ Define whether pathogens are completely heterotypic, partially heterotypic, or homotypic """
-    COMPLETE_HETERO = 1
-    PARTIAL_HETERO = 2
-    HOMOTYPIC = 3
-
-
-class Pathogen(sc.quickobj):
-    """
-    Pathogen dynamics
-    """
-    def __init__(self, sim, is_reassortant, creation_time, is_severe=False, host=None, strain=None):
-        self.sim = sim
-        self.host = host
-        self.creation_time = creation_time
-        self.is_reassortant = is_reassortant
-        self.strain = strain
-        self.is_severe = is_severe
-
-        self.fitness_map = {
-            1: {'default': 1},
-            2: {'default': 0.90, (1, 1): 0.93, (2, 2): 0.93, (3, 3): 0.93, (4, 4): 0.93},
-            3: {'default': 0.87, (1, 1): 0.93, (2, 2): 0.93, (3, 3): 0.90, (4, 4): 0.90},
-            4: {'default': 1, (1, 1): 1, (2, 2): 0.2},
-            5: {'default': 0.2, (1, 1): 1, (2, 1): 0.5, (1, 3): 0.5},
-            6: {'default': 0.05, (1, 8): 1, (2, 4): 0.2, (3, 8): 0.4, (4, 8): 0.5},
-            7: {'default': 0.05, (1, 8): 1, (2, 4): 0.3, (3, 8): 0.7, (4, 8): 0.6},
-            8: {'default': 0.05, (1, 8): 1, (2, 4): 0.4, (3, 8): 0.9, (4, 8): 0.8},
-            9: {'default': 0.2, (1, 8): 1, (2, 4): 0.5, (3, 8): 0.9, (4, 8): 0.8},
-            10: {'default': 0.4, (1, 8): 1, (2, 4): 0.6, (3, 8): 0.9, (4, 8): 0.9},
-            11: {'default': 0.5, (1, 8): 0.98, (2, 4): 0.7, (3, 8): 0.8, (4, 8): 0.8},
-            12: {'default': 0.5, (1, 8): 0.98, (2, 4): 0.8, (3, 8): 0.9, (4, 8): 0.9},
-            13: {'default': 0.7, (1, 8): 0.98, (2, 4): 0.8, (3, 8): 0.9, (4, 8): 0.9},
-            14: {'default': 0.05, (1, 8): 0.98, (2, 4): 0.4, (3, 8): 0.7, (4, 8): 0.6, (9, 8): 0.7, (12, 8): 0.75, (9, 6): 0.58, (11, 8): 0.2},
-            15: {'default': 0.4, (1, 8): 1, (2, 4): 0.7, (3, 8): 0.93, (4, 8): 0.93, (9, 8): 0.95, (12, 8): 0.94, (9, 6): 0.3, (11, 8): 0.35},
-            16: {'default': 0.4, (1, 8): 1, (2, 4): 0.7, (3, 8): 0.85, (4, 8): 0.88, (9, 8): 0.95, (12, 8): 0.93, (9, 6): 0.85, (12, 6): 0.90, (9, 4): 0.90, (1, 6): 0.6, (2, 8): 0.6, (2, 6): 0.6},
-            17: {'default': 0.7, (1, 8): 1, (2, 4): 0.85, (3, 8): 0.85, (4, 8): 0.88, (9, 8): 0.95, (12, 8): 0.93, (9, 6): 0.83, (12, 6): 0.90, (9, 4): 0.90, (1, 6): 0.8, (2, 8): 0.8, (2, 6): 0.8},
-            18: {'default': 0.65, (1, 8): 1, (2, 4): 0.92, (3, 8): 0.79, (4, 8): 0.81, (9, 8): 0.95, (12, 8): 0.89, (9, 6): 0.80, (12, 6): 0.86, (9, 4): 0.83, (1, 6): 0.75, (2, 8): 0.75, (2, 6): 0.75},
-            19: {'default': 0.4, (1, 8): 1, (2, 4): 0.5, (3, 8): 0.55, (4, 8): 0.55, (9, 8): 0.6},
-        }
-        return
-
-    # compares two strains
-    # if they both have the same antigenic segments we return homotypic
-    def match(self, strainIn):
-        numAgSegments = self.sim.numAgSegments
-        if strainIn[:numAgSegments] == self.strain[:numAgSegments]:
-            return PathogenMatch.HOMOTYPIC
-
-        strains_match = False
-        for i in range(numAgSegments):
-            if strainIn[i] == self.strain[i]:
-                strains_match = True
-
-        if strains_match:
-            return PathogenMatch.PARTIAL_HETERO
-        else:
-            return PathogenMatch.COMPLETE_HETERO
-
-    def get_fitness(self):
-        """ Get the fitness based on the fitness hypothesis and the two strains """
-        fitness_hypothesis = self.sim.fitness_hypothesis
-        key = (self.strain[0], self.strain[1])
-
-        if fitness_hypothesis == 1:
-                return 1
-
-        elif fitness_hypothesis == 2:
-            	default = 0.90
-            	mapping = {
-            		(1, 1): 0.93,
-            		(2, 2): 0.93,
-            		(3, 3): 0.93,
-            		(4, 4): 0.93,
-            	}
-            	return mapping.get(key, default)
-
-        elif fitness_hypothesis == 3:
-                default = 0.87
-                mapping = {
-                	(1, 1): 0.93,
-                	(2, 2): 0.93,
-                	(3, 3): 0.90,
-                	(4, 4): 0.90,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 4:
-                default = 1
-                mapping = {
-                	(1, 1): 1,
-                	(2, 2): 0.2,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 5:
-                default = 0.2
-                mapping = {
-                	(1, 1): 1,
-                	(2, 1): 0.5,
-                	(1, 3): 0.5,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 6:
-                default = 0.05
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.2,
-                	(3, 8): 0.4,
-                	(4, 8): 0.5,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 7:
-                default = 0.05
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.3,
-                	(3, 8): 0.7,
-                	(4, 8): 0.6,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 8:
-                default = 0.05
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.4,
-                	(3, 8): 0.9,
-                	(4, 8): 0.8,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 9:
-                default = 0.2
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.6,
-                	(3, 8): 0.9,
-                	(4, 8): 0.9,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 10:
-                default = 0.4
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.6,
-                	(3, 8): 0.9,
-                	(4, 8): 0.9,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 11:
-                default = 0.5
-                mapping = {
-                	(1, 8): 0.98,
-                	(2, 4): 0.7,
-                	(3, 8): 0.8,
-                	(4, 8): 0.8,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 12:
-                default=0.5
-                mapping = {
-                	(1, 8): 0.98,
-                	(2, 4): 0.7,
-                	(3, 8): 0.9,
-                	(4, 8): 0.9,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 13:
-            	default = 0.7
-            	mapping = {
-            		(1, 8): 0.98,
-            		(2, 4): 0.8,
-            		(3, 8): 0.9,
-            		(4, 8): 0.9,
-            	}
-            	return mapping.get(key, default)
-
-        elif fitness_hypothesis == 14:
-                default = 0.05
-                mapping = {
-                	(1, 8): 0.98,
-                	(2, 4): 0.4,
-                	(3, 8): 0.7,
-                	(12, 8): 0.75,
-                	(9, 6): 0.58,
-                	(11, 8): 0.2,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 15:
-                default = 0.4
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.7,
-                	(3, 8): 0.93,
-                	(4, 8): 0.93,
-                	(9, 8): 0.95,
-                	(12, 8): 0.94,
-                	(9, 6): 0.3,
-                	(11, 8): 0.35,
-                }
-                return mapping.get(key, default)
-
-
-        elif fitness_hypothesis == 16:
-                default = 0.4
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.7,
-                	(3, 8): 0.85,
-                	(4, 8): 0.88,
-                	(9, 8): 0.95,
-                	(12, 8): 0.93,
-                	(9, 6): 0.85,
-                	(12, 6): 0.90,
-                	(9, 4): 0.90,
-                	(1, 6): 0.6,
-                	(2, 8): 0.6,
-                	(2, 6): 0.6,
-                }
-                return mapping.get(key, default)
-
-        elif fitness_hypothesis == 17:
-                default = 0.7
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.85,
-                	(3, 8): 0.85,
-                	(4, 8): 0.88,
-                	(9, 8): 0.95,
-                	(12, 8): 0.93,
-                	(9, 6): 0.83,
-                	(12, 6): 0.90,
-                	(9, 4): 0.90,
-                	(1, 6): 0.8,
-                	(2, 8): 0.8,
-                	(2, 6): 0.8,
-                }
-                return mapping.get(key, default)
-
-            # below fitness hypo. 18 was used in the analysis for the high baseline diversity setting in the report
-        elif fitness_hypothesis == 18:
-                default = 0.65
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.92,
-                	(3, 8): 0.79,
-                	(4, 8): 0.81,
-                	(9, 8): 0.95,
-                	(12, 8): 0.89,
-                	(9, 6): 0.80,
-                	(12, 6): 0.86,
-                	(9, 4): 0.83,
-                	(1, 6): 0.75,
-                	(2, 8): 0.75,
-                	(2, 6): 0.75,
-                }
-                return mapping.get(key, default)
-
-            # below fitness hypo 19 was used for the low baseline diversity setting analysis in the report
-        elif fitness_hypothesis == 19:
-                default = 0.4
-                mapping = {
-                	(1, 8): 1,
-                	(2, 4): 0.5,
-                	(3, 8): 0.55,
-                	(4, 8): 0.55,
-                	(9, 8): 0.6,
-                }
-                return mapping.get(key, default)
-
-        else:
-                print("Invalid fitness_hypothesis: ", fitness_hypothesis)
-                exit(-1)
-
-    # def get_fitness(self):
-    #     """ Get the fitness based on the fitness hypothesis and the two strains """
-    #     fitness_hypothesis = self.sim.fitness_hypothesis
-    #     strain = self.strain
-
-    #     if fitness_hypothesis in self.fitness_map:
-    #         fitness_values = self.fitness_map[fitness_hypothesis]
-    #         return fitness_values.get((strain[0], strain[1]), fitness_values.get('default', 1))
-    #     else:
-    #         raise NotImplementedError(f"Invalid fitness_hypothesis: {fitness_hypothesis}")
-
-    def get_strain_name(self):
-        G,P,A,B = [str(self.strain[i]) for i in range(4)]
-        return f'G{G}P{P}A{A}B{B}'
-
-    def __str__(self):
-        return "Strain: " + self.get_strain_name() + " Severe: " + str(self.is_severe) + " Host: " + str(self.host.id) + str(self.creation_time)
 
 
 ### Sim class
-class Sim(sc.quickobj):
+class RotaSim(ss.Sim):
     """
     Run the simulation
     """
 
     def __init__(self,
-            N = 100_000,
+            n_agents = 100_000,
             timelimit = 40,
-            verbose = None,
+            start = 2000,
+            verbose = 0,
             to_csv = True,
             **kwargs,
         ):
@@ -650,6 +94,9 @@ class Sim(sc.quickobj):
             defaults (list): a list of parameters matching the command-line inputs; see below
             verbose (bool): the "verbosity" of the output: if False, print nothing; if None, print the timestep; if True, print out results
         """
+
+        super().__init__(n_agents=n_agents, start=start, stop=start+timelimit, verbose=verbose, **kwargs)
+
         # Define the default parameters
         args = sc.objdict(
             immunity_hypothesis = 1,
@@ -675,8 +122,8 @@ class Sim(sc.quickobj):
         # for i,arg in enumerate(sys.argv[1:]):
         #     args[i] = arg
 
-        if verbose is not False:
-            print(f'Creating simulation with N={N}, timelimit={timelimit} and parameters:')
+        if verbose:
+            print(f'Creating simulation with N={n_agents}, timelimit={timelimit} and parameters:')
             print(args)
 
         # Store parameters directly in the sim
@@ -708,7 +155,7 @@ class Sim(sc.quickobj):
 
         # Set other parameters
         self.to_csv = to_csv # whether to write files
-        self.N = N  # initial population size
+        # self.n_agents = n_agents  # initial population size
         self.timelimit = timelimit  # simulation years
         self.mu = 1.0/70.0     # average life span is 70 years
         self.gamma = 365/7  # 1/average infectious period (1/gamma =7 days)
@@ -728,15 +175,15 @@ class Sim(sc.quickobj):
 
         # Efficacy of the vaccine first dose
         self.vaccine_efficacy_d1 = {
-            PathogenMatch.HOMOTYPIC: 0.6,
-            PathogenMatch.PARTIAL_HETERO: 0.45,
-            PathogenMatch.COMPLETE_HETERO:0.15,
+            rg.PathogenMatch.HOMOTYPIC: 0.6,
+            rg.PathogenMatch.PARTIAL_HETERO: 0.45,
+            rg.PathogenMatch.COMPLETE_HETERO:0.15,
         }
         # Efficacy of the vaccine second dose
         self.vaccine_efficacy_d2 = {
-            PathogenMatch.HOMOTYPIC: 0.8,
-            PathogenMatch.PARTIAL_HETERO: 0.65,
-            PathogenMatch.COMPLETE_HETERO:0.35,
+            rg.PathogenMatch.HOMOTYPIC: 0.8,
+            rg.PathogenMatch.PARTIAL_HETERO: 0.65,
+            rg.PathogenMatch.COMPLETE_HETERO:0.35,
         }
 
         self.vaccination_single_dose_waning_rate = 365/273 #365/1273
@@ -754,7 +201,7 @@ class Sim(sc.quickobj):
         self.reassortment_count = 0
         self.pop_id = 0
         self.t = 0.0
-        self.results = sc.objdict(
+        self.rota_results = sc.objdict(
             columns = ["id", "Strain", "CollectionTime", "Age", "Severity", "InfectionTime", "PopulationSize"],
             infected_all = [],
         )
@@ -1138,158 +585,167 @@ class Sim(sc.quickobj):
                     writer = csv.writer(outputfile)
                     writer.writerows(collected_vaccination_data)
 
-    def prepare_run(self):
+    def init(self, force=False):
         """
         Set up the variables for the run
         """
-        # relative protection for infection from natural immunity
-        immunity_hypothesis = self.immunity_hypothesis
+        if force or not self.initialized:
 
-        # Define the mapping of immunity_hypothesis to their corresponding rates using tuples
-        immunity_rates = {
-            1: (0, 0, 0),
-            2: (0, 1, 0),
-            3: (0, 0.5, 0),
-            4: (0, 0.95, 0.9),
-            5: (0, 0.95, 0.90),
-            7: (0.95, 0.90, 0.2),
-            8: (0.9, 0.5, 0),
-            9: (0.9, 0.45, 0.35),
-            10: (0.8, 0.45, 0.35),
-        }
+            if self.pars.people is None:
+                self.pars.people = ss.People(n_agents=self.pars.n_agents)
 
-        # Get the rates for the given immunity_hypothesis
-        rates = immunity_rates.get(immunity_hypothesis)
+            super().init(force=force)
 
-        if rates is None:
-            raise NotImplementedError(f"No partial cross immunity rate for immunity hypothesis: {immunity_hypothesis}")
+            # relative protection for infection from natural immunity
+            immunity_hypothesis = self.immunity_hypothesis
 
-        # Unpack the tuple into the corresponding variables
-        homotypic_immunity_rate, partial_cross_immunity_rate, complete_heterotypic_immunity_rate = rates
+            # Define the mapping of immunity_hypothesis to their corresponding rates using tuples
+            immunity_rates = {
+                1: (0, 0, 0),
+                2: (0, 1, 0),
+                3: (0, 0.5, 0),
+                4: (0, 0.95, 0.9),
+                5: (0, 0.95, 0.90),
+                7: (0.95, 0.90, 0.2),
+                8: (0.9, 0.5, 0),
+                9: (0.9, 0.45, 0.35),
+                10: (0.8, 0.45, 0.35),
+            }
 
-        self.homotypic_immunity_rate = homotypic_immunity_rate
-        self.partial_cross_immunity_rate = partial_cross_immunity_rate
-        self.complete_heterotypic_immunity_rate = complete_heterotypic_immunity_rate
+            # Get the rates for the given immunity_hypothesis
+            rates = immunity_rates.get(immunity_hypothesis)
 
-        self.done_vaccinated = False
+            if rates is None:
+                raise NotImplementedError(f"No partial cross immunity rate for immunity hypothesis: {immunity_hypothesis}")
 
-        self.vaccine_efficacy_i_d1 = {}
-        self.vaccine_efficacy_s_d1 = {}
-        self.vaccine_efficacy_i_d2 = {}
-        self.vaccine_efficacy_s_d2 = {}
-        for (k, v) in self.vaccine_efficacy_d1.items():
-            (ve_i, ve_s) = self.breakdown_vaccine_efficacy(v, self.ve_i_to_ve_s_ratio)
-            self.vaccine_efficacy_i_d1[k] = ve_i
-            self.vaccine_efficacy_s_d1[k] = ve_s
-        for (k, v) in self.vaccine_efficacy_d2.items():
-            (ve_i, ve_s) = self.breakdown_vaccine_efficacy(v, self.ve_i_to_ve_s_ratio)
-            self.vaccine_efficacy_i_d2[k] = ve_i
-            self.vaccine_efficacy_s_d2[k] = ve_s
+            # Unpack the tuple into the corresponding variables
+            homotypic_immunity_rate, partial_cross_immunity_rate, complete_heterotypic_immunity_rate = rates
 
-        if self.verbose: print("VE_i: ", self.vaccine_efficacy_i_d1)
-        if self.verbose: print("VE_s: ", self.vaccine_efficacy_s_d1)
+            self.homotypic_immunity_rate = homotypic_immunity_rate
+            self.partial_cross_immunity_rate = partial_cross_immunity_rate
+            self.complete_heterotypic_immunity_rate = complete_heterotypic_immunity_rate
 
-        # Vaccination rates are derived based on the following formula
-        self.vaccine_second_dose_rate = 0.8
-        self.vaccine_first_dose_rate = math.sqrt(self.vaccine_second_dose_rate)
-        if self.verbose: print("Vaccination - first dose rate: %s, second dose rate %s" % (self.vaccine_first_dose_rate, self.vaccine_second_dose_rate))
+            self.done_vaccinated = False
 
-        self.total_strain_counts_vaccine = {}
+            self.vaccine_efficacy_i_d1 = {}
+            self.vaccine_efficacy_s_d1 = {}
+            self.vaccine_efficacy_i_d2 = {}
+            self.vaccine_efficacy_s_d2 = {}
+            for (k, v) in self.vaccine_efficacy_d1.items():
+                (ve_i, ve_s) = self.breakdown_vaccine_efficacy(v, self.ve_i_to_ve_s_ratio)
+                self.vaccine_efficacy_i_d1[k] = ve_i
+                self.vaccine_efficacy_s_d1[k] = ve_s
+            for (k, v) in self.vaccine_efficacy_d2.items():
+                (ve_i, ve_s) = self.breakdown_vaccine_efficacy(v, self.ve_i_to_ve_s_ratio)
+                self.vaccine_efficacy_i_d2[k] = ve_i
+                self.vaccine_efficacy_s_d2[k] = ve_s
 
-        numSegments = 4
-        numNoneAgSegments = 2
-        self.numAgSegments = numSegments - numNoneAgSegments
-        #segmentVariants = [[i for i in range(1, 3)], [i for i in range(1, 3)], [i for i in range(1, 2)], [i for i in range(1, 2)]]     ## creating variats for the segments
-        segmentVariants = [[1,2,3,4,9,11,12], [8,4,6], [i for i in range(1, 2)], [i for i in range(1, 2)]]
-        # segmentVariants for the Low baseline diversity setting
-        #segmentVariants = [[1,2,3,4,9], [8,4], [i for i in range(1, 2)], [i for i in range(1, 2)]]
-        segment_combinations = [tuple(i) for i in itertools.product(*segmentVariants)]  # getting all possible combinations from a list of list
-        rnd.shuffle(segment_combinations)
-        number_all_strains = len(segment_combinations)
-        n_init_seg = 100
-        initial_segment_combinations = {
-            (1,8,1,1) : n_init_seg,
-            (2,4,1,1) : n_init_seg,
-            (9,8,1,1) : n_init_seg,
-            (4,8,1,1) : n_init_seg,
-            (3,8,1,1) : n_init_seg,
-            (12,8,1,1): n_init_seg,
-            (12,6,1,1): n_init_seg,
-            (9,4,1,1) : n_init_seg,
-            (9,6,1,1) : n_init_seg,
-            (1,6,1,1) : n_init_seg,
-            (2,8,1,1) : n_init_seg,
-            (2,6,1,1) : n_init_seg,
-            (11,8,1,1): n_init_seg,
-            (11,6,1,1): n_init_seg,
-            (1,4,1,1) : n_init_seg,
-            (12,4,1,1): n_init_seg,
-        }
-        # initial strains for the Low baseline diversity setting
-        #initial_segment_combinations = {(1,8,1,1): 100, (2,4,1,1): 100} #, (9,8,1,1): 100} #, (4,8,1,1): 100}
+            if self.verbose: print("VE_i: ", self.vaccine_efficacy_i_d1)
+            if self.verbose: print("VE_s: ", self.vaccine_efficacy_s_d1)
 
-        # Track the number of immune hosts(immunity_counts) in the host population
-        infected_pop = []
-        pathogens_pop = []
+            # Vaccination rates are derived based on the following formula
+            self.vaccine_second_dose_rate = 0.8
+            self.vaccine_first_dose_rate = math.sqrt(self.vaccine_second_dose_rate)
+            if self.verbose: print("Vaccination - first dose rate: %s, second dose rate %s" % (self.vaccine_first_dose_rate, self.vaccine_second_dose_rate))
 
-        # for each strain track the number of hosts infected with it at current time: strain_count
-        strain_count = {}
+            self.total_strain_counts_vaccine = {}
 
-        # for each number in range of N, make a new Host object, i is the id.
-        host_pop = HostPop(self.N, self)
+            numSegments = 4
+            numNoneAgSegments = 2
+            self.numAgSegments = numSegments - numNoneAgSegments
+            #segmentVariants = [[i for i in range(1, 3)], [i for i in range(1, 3)], [i for i in range(1, 2)], [i for i in range(1, 2)]]     ## creating variats for the segments
+            segmentVariants = [[1,2,3,4,9,11,12], [8,4,6], [i for i in range(1, 2)], [i for i in range(1, 2)]]
+            # segmentVariants for the Low baseline diversity setting
+            #segmentVariants = [[1,2,3,4,9], [8,4], [i for i in range(1, 2)], [i for i in range(1, 2)]]
+            segment_combinations = [tuple(i) for i in itertools.product(*segmentVariants)]  # getting all possible combinations from a list of list
+            rnd.shuffle(segment_combinations)
+            number_all_strains = len(segment_combinations)
+            n_init_seg = 100
+            initial_segment_combinations = {
+                (1,8,1,1) : n_init_seg,
+                (2,4,1,1) : n_init_seg,
+                (9,8,1,1) : n_init_seg,
+                (4,8,1,1) : n_init_seg,
+                (3,8,1,1) : n_init_seg,
+                (12,8,1,1): n_init_seg,
+                (12,6,1,1): n_init_seg,
+                (9,4,1,1) : n_init_seg,
+                (9,6,1,1) : n_init_seg,
+                (1,6,1,1) : n_init_seg,
+                (2,8,1,1) : n_init_seg,
+                (2,6,1,1) : n_init_seg,
+                (11,8,1,1): n_init_seg,
+                (11,6,1,1): n_init_seg,
+                (1,4,1,1) : n_init_seg,
+                (12,4,1,1): n_init_seg,
+            }
+            # initial strains for the Low baseline diversity setting
+            #initial_segment_combinations = {(1,8,1,1): 100, (2,4,1,1): 100} #, (9,8,1,1): 100} #, (4,8,1,1): 100}
 
-        self.pop_id = self.N
-        self.to_be_vaccinated_pop = []
-        self.single_dose_vaccinated_pop = []
+            # Track the number of immune hosts(immunity_counts) in the host population
+            infected_pop = []
+            pathogens_pop = []
 
-        # Store these for later
-        self.infected_pop = infected_pop
-        self.pathogens_pop = pathogens_pop
-        self.host_pop = host_pop
-        self.strain_count = strain_count
+            # for each strain track the number of hosts infected with it at current time: strain_count
+            strain_count = {}
 
-        for i in range(number_all_strains):
-            self.strain_count[segment_combinations[i]] = 0
+            # for each number in range of N, make a new Host object, i is the id.
+            host_pop = ss.People(self.n_agents)
 
-        # if initial immunity is true
-        if self.verbose:
-            if self.initial_immunity:
-                print("Initial immunity is set to True")
-            else:
-                print("Initial immunity is set to False")
+            self.pop_id = self.n_agents
+            self.to_be_vaccinated_pop = []
+            self.single_dose_vaccinated_pop = []
 
-        ### infecting the initial infecteds
-        for (initial_strain, num_infected) in initial_segment_combinations.items():
-            if self.initial_immunity:
-                for j in range(self.num_initial_immune):
+            # Store these for later
+            self.infected_pop = infected_pop
+            self.pathogens_pop = pathogens_pop
+            self.host_pop = host_pop
+            self.strain_count = strain_count
+
+            for i in range(number_all_strains):
+                self.strain_count[segment_combinations[i]] = 0
+
+            # if initial immunity is true
+            if self.verbose:
+                if self.initial_immunity:
+                    print("Initial immunity is set to True")
+                else:
+                    print("Initial immunity is set to False")
+
+            ### infecting the initial infecteds
+            for (initial_strain, num_infected) in initial_segment_combinations.items():
+                if self.initial_immunity:
+                    for j in range(self.num_initial_immune):
+                        h = rnd.choice(host_pop)
+                        h.immunity[initial_strain] = self.t
+                        self.immunity_counts += 1
+                        h.is_immune_flag = True
+
+                for j in range(num_infected):
                     h = rnd.choice(host_pop)
-                    h.immunity[initial_strain] = self.t
-                    self.immunity_counts += 1
-                    h.is_immune_flag = True
+                    if not h.isInfected():
+                        infected_pop.append(h)
+                    p = rg.RotaPathogen(self, False, self.t, host = h, strain = initial_strain)
+                    pathogens_pop.append(p)
+                    h.infecting_pathogen.append(p)
+                    strain_count[p.strain] += 1
+            if self.verbose: print(strain_count)
 
-            for j in range(num_infected):
-                h = rnd.choice(host_pop)
-                if not h.isInfected():
-                    infected_pop.append(h)
-                p = Pathogen(self, False, self.t, host = h, strain = initial_strain)
-                pathogens_pop.append(p)
-                h.infecting_pathogen.append(p)
-                strain_count[p.strain] += 1
-        if self.verbose: print(strain_count)
+            if self.to_csv:
+                self.initialize_files(strain_count)
 
-        if self.to_csv:
-            self.initialize_files(strain_count)
+            self.tau_steps = 0
+            self.last_data_colllected = 0
+            self.data_collection_rate = 0.1
 
-        self.tau_steps = 0
-        self.last_data_colllected = 0
-        self.data_collection_rate = 0.1
+            for strain, count in strain_count.items():
+                if strain[:self.numAgSegments] in self.total_strain_counts_vaccine:
+                    self.total_strain_counts_vaccine[strain[:self.numAgSegments]] += count
+                else:
+                    self.total_strain_counts_vaccine[strain[:self.numAgSegments]] = count
 
-        for strain, count in strain_count.items():
-            if strain[:self.numAgSegments] in self.total_strain_counts_vaccine:
-                self.total_strain_counts_vaccine[strain[:self.numAgSegments]] += count
-            else:
-                self.total_strain_counts_vaccine[strain[:self.numAgSegments]] = count
-        return
+
+        return self
 
     def integrate(self):
         """
@@ -1432,7 +888,7 @@ class Sim(sc.quickobj):
         """
         Run the simulation
         """
-        self.prepare_run()
+        self.init()
         events = self.integrate()
         self.to_df()
         return events
