+++ tests/simple.py
@@ -2,6 +2,6 @@ import sciris as sc
 import rotasim as rs
 
 with sc.timer():
-    sim = rs.Sim()
+    sim = rs.RotaSim()
     events = sim.run()
     print(events)
